Integrate MudBlazor for enhanced UI and functionality
- Added MudBlazor package reference in `CDS.SQLiteLogViewer.csproj`.
- Updated `MainLayout.razor` to use MudBlazor components, improving layout with a responsive design and sidebar navigation.
- Changed `DbId` type from `int` to `long` in `LogEntry.cs` for larger identifier support.
- Replaced HTML table with `MudTable` in `LogViewer.razor`, adding search functionality and improved file upload button.
- Included MudBlazor services in `Program.cs` and linked CSS/JS files in `index.html` for proper component styling and functionality.
- Ensured consistency in `LogReader.cs` with updated `DbId` type.

+++ CDS.SQLiteLogViewer/Models/LogEntry.cs
@@ -2,7 +2,7 @@
 
 public class LogEntry
 {
-    public int DbId { get; set; }
+    public long DbId { get; set; }
     public LogLevel Level { get; set; }
     public string? Category { get; set; }
     public int EventId { get; set; }
